# openapi.yaml
openapi: 3.0.3
info:
  title: Kinetic Daemon API
  version: 0.1.0
  description: The HTTP REST API exposed by the Kinetic Daemon for domain management, VDF proofs, and resolution.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: "{scheme}://{host}:{port}/api"
    description: Kinetic daemon API — configure host and port at runtime
    variables:
      scheme:
        default: http
        enum: [http, https]
        description: Transport scheme (http for local, https for remote/TLS deployments)
      host:
        default: "127.0.0.1"
        description: Daemon bind address — override per deployment
      port:
        default: "16002"
        description: Daemon API port — override per deployment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    PublishResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required:
        - status
        - message
...
    VdfProof:
      type: object
      properties:
        proof_bytes:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 255
      required:
        - proof_bytes
    PreviousProof:
      type: object
      properties:
        salt:
          type: array
          items:
            type: integer
        kyn:
          type: integer
        drand_signature:
          type: string
        iterations:
          type: integer
        vdf_proof:
          $ref: '#/components/schemas/VdfProof'
        signature:
          type: array
          items:
            type: integer
      required:
        - salt
        - kyn
        - drand_signature
        - iterations
        - vdf_proof
        - signature
    Reveal:
      type: object
      properties:
        protocol_version:
          type: integer
          default: 2
        name:
          type: string
        payload:
          type: array
          items:
            type: integer
        salt:
          type: array
          items:
            type: integer
        kyn:
          type: integer
        drand_signature:
          type: string
        iterations:
          type: integer
        vdf_proof:
          $ref: '#/components/schemas/VdfProof'
        pubkey:
          type: array
          items:
            type: integer
        signature:
          type: array
          items:
            type: integer
        previous_proof:
          $ref: '#/components/schemas/PreviousProof'
          nullable: true
        miner_pubkey:
          type: array
          items:
            type: integer
          nullable: true
      required:
        - protocol_version
        - name
        - payload
        - salt
        - kyn
        - drand_signature
        - iterations
        - vdf_proof
        - pubkey
        - signature
    Commitment:
      type: object
      properties:
        hash:
          type: array
          items:
            type: integer
      required:
        - hash
    CommitRequest:
      type: object
      properties:
        name:
          type: string
        commitment:
          $ref: '#/components/schemas/Commitment'
      required:
        - name
        - commitment
    PublishRequest:
      type: object
      properties:
        reveal:
          $ref: '#/components/schemas/Reveal'
      required:
        - reveal
    VdfRegisterRequest:
      type: object
      properties:
        name:
          type: string
        iterations:
          type: integer
          nullable: true
      required:
        - name
...
    KidDocument:
      type: object
      properties:
        doc_type:
          type: string
        kid:
          type: string
        created_at:
          type: integer
        controller_keys:
          type: array
          items:
            type: string
        manifest:
          type: object
          nullable: true
        revocation_keys:
          type: array
          items:
            type: string
        signature:
          type: string
          nullable: true
      required:
        - doc_type
        - kid
        - created_at
        - controller_keys
        - revocation_keys
    AuthorizedKid:
      type: object
      properties:
        name:
          type: string
        kid_doc:
          $ref: '#/components/schemas/KidDocument'
        owner_signature:
          type: array
          items:
            type: integer
      required:
        - name
        - kid_doc
        - owner_signature
...
paths:
...
  /api/v1/micro/nrs/record/commit:
    post:
      operationId: commitName
      summary: Commit a name hash to DHT [Private]
      description: Requires Nrs or Admin role.
      tags: [NRS]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
      responses:
        '200':
          description: Committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
  /api/v1/micro/nrs/record/publish:
    post:
      operationId: publishName
      summary: Publish a name reveal to DHT [Private]
      description: Requires Nrs or Admin role.
      tags: [NRS]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishRequest'
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
  /api/v1/micro/kid/publish:
    post:
      operationId: publishKid
      summary: Publish a KID to DHT [Private]
      description: Requires Kid or Admin role.
      tags: [KID]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizedKid'
      responses:
        '200':
          description: Published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
...
  /api/v1/macro/register:
    post:
      operationId: vdfRegister
      summary: Start VDF name registration task [Private]
      description: Requires VDF or Admin role.
      tags: [VDF]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VdfRegisterRequest'
      responses:
        '200':
          description: Task started
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  message:
                    type: string
...